# core/docs/openapi.yaml
openapi: 3.1.0
info:
  title: CauseFlow API
  description: |
    CauseFlow — AI SRE platform. REST API for incident ingestion, triage,
    investigation, remediation, memory/knowledge, integrations, billing,
    tenant/user/team management, notifications, audit, widget/portal, and
    skills.

    Architecture: Modular monolith (Modlito) + Clean Architecture.
    Stack: TypeScript + Hono + ElectroDB (DynamoDB single-table) + Redis + SQS.

    ## Authentication
    Most dashboard endpoints use `bearerAuth` (Clerk JWT).
    Ingestion webhook endpoints accept either a shared webhook secret or a
    tenant API key (`X-API-Key`). Widget endpoints use `X-API-Key`.
    Stripe, Clerk and Composio webhooks verify provider signatures and are
    mounted as public.
  version: 0.1.0
  contact:
    name: CauseFlow
servers:
  - url: https://api.causeflow.io
    description: Production
  - url: https://staging-api.causeflow.io
    description: Staging
  - url: http://localhost:3000
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /v1/tenants:
    post:
      tags: [Tenant]
      summary: Create a tenant (admin/owner)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateTenantInput' }
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    get:
      tags: [Tenant]
      summary: List tenants owned by the authenticated user
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Cursor'
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Tenant' } }
                  cursor: { type: string, nullable: true }
  /v1/tenants/{tenantId}:
    get:
      tags: [Tenant]
      summary: Get a tenant by id
      parameters: [ { $ref: '#/components/parameters/TenantIdPath' } ]
      responses:
        '200':
          description: Tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [Tenant]
      summary: Update a tenant (admin/owner)
      parameters: [ { $ref: '#/components/parameters/TenantIdPath' } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/UpdateTenantInput' }
      responses:
        '200':
          description: Updated tenant
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Tenant' }

  # ─────────────────────────── User ───────────────────────────
...
  /v1/memory/chat:
    post:
      tags: [Memory]
      summary: Unified chat (Hindsight/router decides intent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [message]
              properties:
                message: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '200':
          description: Chat response
          content:
            application/json:
              schema: { type: object, additionalProperties: true }
        '503': { description: Chat not available }
...
  /v1/memory/chat/history:
    get:
      tags: [Memory]
      summary: List recent chats
      parameters:
        - { in: query, name: limit, schema: { type: integer, default: 20, minimum: 1, maximum: 100 } }
      responses:
        '200':
          description: Chats
          content:
            application/json:
              schema:
                type: object
                properties:
                  chats: { type: array, items: { type: object, additionalProperties: true } }
        '503': { description: Chat history not available }
  /v1/memory/chat/history/{chatId}:
    get:
      tags: [Memory]
      summary: Get messages of a chat
      parameters:
        - { in: path, name: chatId, required: true, schema: { type: string } }
      responses:
        '200':
          description: Messages
          content:
            application/json:
              schema:
                type: object
                properties:
                  chatId: { type: string }
                  messages: { type: array, items: { type: object, additionalProperties: true } }
        '503': { description: Chat history not available }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Clerk-issued JWT. Tenant context derived from `o.id` claim.
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Tenant-scoped API key for programmatic and widget access.
    webhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: Shared webhook secret for ingestion endpoints.
    webhookSignature:
      type: apiKey
      in: header
      name: webhook-signature
      description: HMAC signature header used by Composio webhooks.
    stripeSignature:
      type: apiKey
      in: header
      name: stripe-signature
      description: Stripe webhook signature header.
    svixSignature:
      type: apiKey
      in: header
      name: svix-signature
      description: Svix signature for Clerk webhooks.
  parameters:
    Limit:
      in: query
      name: limit
      schema: { type: integer, default: 20, minimum: 1, maximum: 200 }
    Cursor:
      in: query
      name: cursor
      schema: { type: string }
    TenantIdPath:
      in: path
      name: tenantId
      required: true
      schema: { type: string }
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Forbidden:
      description: Caller lacks required role
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
        message: { type: string }
        details: { type: object, additionalProperties: true }
    Tenant:
      type: object
      properties:
        tenantId: { type: string }
        name: { type: string }
        slug: { type: string }
        ownerEmail: { type: string, format: email }
        plan: { type: string, enum: [starter, pro, business, enterprise] }
        status: { type: string, enum: [active, suspended, trial, cancelled] }
        settings:
          type: object
          properties:
            maxIncidentsPerMonth: { type: integer }
            autoRemediation: { type: boolean }
            notificationChannels: { type: array, items: { type: string } }
            chatProvider: { type: string, enum: [web_portal, slack, teams] }
            widgetConfig: { type: object, additionalProperties: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateTenantInput:
      type: object
      required: [name, slug, ownerEmail]
      properties:
        name: { type: string, minLength: 2, maxLength: 100 }
        slug: { type: string, minLength: 2, maxLength: 50, pattern: '^[a-z0-9-]+$' }
        ownerEmail: { type: string, format: email }
        plan: { type: string, enum: [starter, pro, business, enterprise] }
        settings:
          type: object
          properties:
            maxIncidentsPerMonth: { type: integer }
            autoRemediation: { type: boolean }
            notificationChannels: { type: array, items: { type: string } }
    UpdateTenantInput:
      type: object
      properties:
        name: { type: string }
        ownerEmail: { type: string, format: email }
        plan: { type: string, enum: [starter, pro, business, enterprise] }
        status: { type: string, enum: [active, suspended, trial, cancelled] }
        settings:
          type: object
          properties:
            maxIncidentsPerMonth: { type: integer }
            autoRemediation: { type: boolean }
            notificationChannels: { type: array, items: { type: string } }
            chatProvider: { type: string, enum: [web_portal, slack, teams] }
...